requirements.txt
streamlit
streamlit-authenticator
PyYAML
biopython
pandas
sqlite3